#21 is now ready for the test.
/accounts/urls.py now has a url for operator managment (from normal admin)
/sales/add_sale and /sales/edit_sale has changed for show only related shops to operator.
/sales/ListSalesView only shows related sales to operators and admins.

+++ backtoshops/shops/views.py
@@ -1,10 +1,10 @@
 import json
 import settings
 from django.core.urlresolvers import reverse
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseRedirect
 from django.views.generic.base import TemplateView
 from django.views.generic.edit import CreateView, DeleteView, UpdateView
-from fouillis.views import BOLoginRequiredMixin
+from fouillis.views import LoginRequiredMixin
 from shops.models import Shop
 from shops.forms import ShopForm
 
@@ -18,7 +18,7 @@ class ShopCoordinates(TemplateView):
             toret['longitude'] = shop.longitude
         return HttpResponse(content=json.dumps(toret), mimetype="application/json")
 
-class BaseShopView(BOLoginRequiredMixin):
+class BaseShopView(LoginRequiredMixin):
     template_name = "physical_shop.html"
     form_class = ShopForm
     model = Shop
@@ -27,9 +27,12 @@ class BaseShopView(BOLoginRequiredMixin):
         return reverse("page_shops")
 
     def get_context_data(self, **kwargs):
+        shops = Shop.objects.filter(mother_brand=self.request.user.get_profile().work_for)
+        if not self.request.user.is_staff: #operator
+            shops = shops.filter(pk__in=self.request.user.get_profile().shops.all())
         kwargs.update({
             'shop_pk': self.kwargs.get('pk', None),
-            'shops': Shop.objects.filter(mother_brand=self.request.user.get_profile().work_for),
+            'shops': shops,
             'request': self.request,
             'geonames_username': settings.GEONAMES_USERNAME,
             'media_url': settings.MEDIA_URL,
@@ -41,6 +44,12 @@ class CreateShopView(BaseShopView, CreateView):
         return {
             "mother_brand": self.request.user.get_profile().work_for
         }
+    
+    def post(self, request, *args, **kwargs):
+        if request.user.is_staff:
+            return super(CreateShopView,self).post(request, *args, **kwargs)
+        else:
+            return HttpResponseRedirect('/')    
         
     def get_success_url(self):
         new_id = Shop.objects.all().count()
@@ -49,16 +58,31 @@ class CreateShopView(BaseShopView, CreateView):
 
 class EditShopView(BaseShopView, UpdateView):
     queryset = Shop.objects.all()
+
+    def get(self, request, *args, **kwargs):
+        pk = self.kwargs.get('pk',None)
+        if request.user.is_staff or len(request.user.get_profile().shops.filter(pk=pk))>0:
+            return super(EditShopView,self).get(request, *args, **kwargs)
+        else:
+            return HttpResponseRedirect('/')
     
+    def post(self, request, *args, **kwargs):
+        pk = self.kwargs.get('pk',None)
+        if request.user.is_staff or len(request.user.get_profile().shops.filter(pk=pk))>0:
+            return super(EditShopView,self).post(request, *args, **kwargs)
+        else:
+            return HttpResponseRedirect('/')
+        
     def get_success_url(self):
         pk = self.kwargs.get('pk', None)
         return reverse("edit_shop",args=[pk])
 
 class DeleteShopView(BaseShopView, DeleteView):
     def delete(self, request, *args, **kwargs):
-        self.object = self.get_object()
-        self.object.delete()
-        return HttpResponse(content=json.dumps({"shop_pk": self.kwargs.get('pk', None)}),
-                            mimetype="application/json")
-
-    
+        if request.user.is_staff:
+            self.object = self.get_object()
+            self.object.delete()
+            return HttpResponse(content=json.dumps({"shop_pk": self.kwargs.get('pk', None)}),
+                                mimetype="application/json")
+        else:
+            return HttpResponse(content=json.dumps({"shop_pk": "N/A"}), mimetype='application/json')
